TRAVIS_Merged/requirements.txt
impyla
mysql-connector==2.2.9
numpy
pandas
paramiko
pure-sasl
pyhive
pymongo
pyodbc
pyyaml
snowflake-connector-python==2.7.12
xlwt
openpyxl
pypdf2
ttkbootstrap
psutil
flatten_json
ijson
pypiwin32
pillow==9.0.0
